+++ src/Contexts/Backoffice/Courses/domain/BackofficeCourseDuration.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+export class BackofficeCourseDuration extends StringValueObject {}
