+++ src/Contexts/Backoffice/Courses/domain/BackofficeCourseId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../../../Shared/domain/value-object/Uuid';
+
+export class BackofficeCourseId extends Uuid {}
