+++ src/Contexts/Backoffice/Courses/domain/BackofficeCourseName.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+export class BackofficeCourseName extends StringValueObject {}
