+++ src/Contexts/Backoffice/SearchAll/application/CoursesFinder.ts
@@ -1,0 +1,12 @@
+import { BackofficeCourseRepository } from '../../Courses/domain/BackofficeCourseRepository';
+import { SearchAllCoursesResponse } from './SearchAllCoursesResponse';
+
+export class CoursesFinder {
+  constructor(private coursesRepository: BackofficeCourseRepository) {}
+
+  async run() {
+    const courses = await this.coursesRepository.searchAll();
+
+    return new SearchAllCoursesResponse(courses);
+  }
+}
