+++ src/Contexts/Backoffice/SearchAll/application/SearchAllCoursesQueryHandler.ts
