+++ src/Contexts/Backoffice/SearchAll/application/SearchAllCoursesResponse.ts
@@ -1,0 +1,9 @@
+import { BackofficeCourse } from '../../Courses/domain/BackofficeCourse';
+
+export class SearchAllCoursesResponse {
+  readonly courses: Array<BackofficeCourse>;
+
+  constructor(courses: Array<BackofficeCourse>) {
+    this.courses = courses;
+  }
+}
