--- src/Contexts/Backoffice/application/SearchAll/SearchAllCoursesResponse.ts
@@ -1,9 +1,0 @@
-import { BackofficeCourse } from '../../domain/BackofficeCourse';
-
-export class SearchAllCoursesResponse {
-  readonly courses: Array<BackofficeCourse>;
-
-  constructor(courses: Array<BackofficeCourse>) {
-    this.courses = courses;
-  }
-}
