--- src/Contexts/Backoffice/domain/BackofficeCourseDuration.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class BackofficeCourseDuration extends StringValueObject {}
