--- src/Contexts/Backoffice/domain/BackofficeCourseId.ts
@@ -1,3 +1,0 @@
-import { Uuid } from '../../Shared/domain/value-object/Uuid';
-
-export class BackofficeCourseId extends Uuid {}
