--- src/Contexts/Backoffice/domain/BackofficeCourseName.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class BackofficeCourseName extends StringValueObject {}
