+++ src/Contexts/Shared/domain/AggregateRoot.ts
@@ -1,0 +1,19 @@
+import { DomainEvent } from './DomainEvent';
+
+export abstract class AggregateRoot {
+  private domainEvents: Array<DomainEvent>;
+
+  constructor() {
+    this.domainEvents = [];
+  }
+
+  pullDomainEvents(): Array<DomainEvent> {
+    return this.domainEvents;
+  }
+
+  record(event: DomainEvent): void {
+    this.domainEvents.push(event);
+  }
+
+  abstract toPrimitives(): any;
+}
